Adds EventsController
